+++ airtable/auth.py
@@ -23,22 +23,15 @@ import requests
 
 
 class AirtableAuth(requests.auth.AuthBase):
-    def __init__(self, api_key=None):
+    def __init__(self, api_key):
         """
         Authentication used by Airtable Class
 
         Args:
-            api_key (``str``): Airtable API Key. Optional.
-                If not set, it will look for
-                enviroment variable ``AIRTABLE_API_KEY``
+            api_key (``str``): Airtable API Key.
         """
         try:
-            self.api_key = api_key or os.environ["AIRTABLE_API_KEY"]
-        except KeyError:
-            raise KeyError(
-                "Api Key not found. Pass api_key as a kwarg \
-                            or set an env var AIRTABLE_API_KEY with your key"
-            )
+            self.api_key = api_key
 
     def __call__(self, request):
         auth_token = {"Authorization": "Bearer {}".format(self.api_key)}